feat(crons): Make a project level `MonitorDetails` api.
We want to make cron slugs unique per project instead of org. To support this, we need project level apis that the frontend can call.

Related to https://github.com/getsentry/sentry/issues/65898
