+++ src/sentry/monitors/endpoints/project_monitors_details.py
@@ -15,7 +15,7 @@ from sentry.apidocs.constants import (
     RESPONSE_UNAUTHORIZED,
 )
 from sentry.apidocs.parameters import GlobalParams, MonitorParams
-from sentry.monitors.endpoints.base import MonitorEndpoint
+from sentry.monitors.endpoints.base import ProjectMonitorEndpoint
 from sentry.monitors.endpoints.monitor_details import MonitorDetailsMixin
 from sentry.monitors.serializers import MonitorSerializer
 from sentry.monitors.validators import MonitorValidator
@@ -24,7 +24,7 @@ from sentry.utils.auth import AuthenticatedHttpRequest
 
 @region_silo_endpoint
 @extend_schema(tags=["Crons"])
-class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
+class ProjectMonitorDetailsEndpoint(ProjectMonitorEndpoint, MonitorDetailsMixin):
     publish_status = {
         "DELETE": ApiPublishStatus.PUBLIC,
         "GET": ApiPublishStatus.PUBLIC,
@@ -36,8 +36,8 @@ class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
         operation_id="Retrieve a Monitor",
         parameters=[
             GlobalParams.ORG_SLUG,
+            GlobalParams.PROJECT_SLUG,
             MonitorParams.MONITOR_SLUG,
-            GlobalParams.ENVIRONMENT,
         ],
         responses={
             200: MonitorSerializer,
@@ -46,7 +46,7 @@ class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
             404: RESPONSE_NOT_FOUND,
         },
     )
-    def get(self, request: Request, organization, project, monitor) -> Response:
+    def get(self, request: Request, project, monitor) -> Response:
         """
         Retrieves details for a monitor.
         """
@@ -56,6 +56,7 @@ class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
         operation_id="Update a Monitor",
         parameters=[
             GlobalParams.ORG_SLUG,
+            GlobalParams.PROJECT_SLUG,
             MonitorParams.MONITOR_SLUG,
         ],
         request=MonitorValidator,
@@ -67,7 +68,7 @@ class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
             404: RESPONSE_NOT_FOUND,
         },
     )
-    def put(self, request: AuthenticatedHttpRequest, organization, project, monitor) -> Response:
+    def put(self, request: AuthenticatedHttpRequest, project, monitor) -> Response:
         """
         Update a monitor.
         """
@@ -77,6 +78,7 @@ class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
         operation_id="Delete a Monitor or Monitor Environments",
         parameters=[
             GlobalParams.ORG_SLUG,
+            GlobalParams.PROJECT_SLUG,
             MonitorParams.MONITOR_SLUG,
             GlobalParams.ENVIRONMENT,
         ],
@@ -88,7 +90,7 @@ class OrganizationMonitorDetailsEndpoint(MonitorEndpoint, MonitorDetailsMixin):
             404: RESPONSE_NOT_FOUND,
         },
     )
-    def delete(self, request: Request, organization, project, monitor) -> Response:
+    def delete(self, request: Request, project, monitor) -> Response:
         """
         Delete a monitor or monitor environments.
         """
